HX711: Cast reading to float, increase initial timeout

+++ hx711.h
@@ -10,6 +10,6 @@
 void hx711_init(TIM_TypeDef *timer);
 int32_t hx711_single_read(void);
 int32_t hx711_avg_read(uint8_t samples);
-int32_t hx711_read_gram(void);
+float hx711_read_gram(void);
 
 #endif // __HX711_H__
